pyproject.toml
[project]
name = "churn-prediction"
version = "0.1.0"
description = "Churn prediction for AvePoint ML exercise"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "lightgbm>=4.0.0",
    "shap>=0.43.0",
    "plotly>=5.18.0",
    "pyarrow>=14.0.0",
    "pyyaml>=6.0.0",
    "mlflow>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "marimo>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "marimo>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]